pyproject.toml
[project]
name = "req-agent"
version = "0.1.0"
description = "Local requirements intake agent (JSON + DOCX/MD) with configurable LLM provider."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-docx>=1.1.0",
  "jsonschema>=4.21",
  "redis>=5.0",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "pyyaml>=6.0.1",
  "streamlit>=1.32",
]

[tool.pytest.ini_options]
pythonpath = ["."]
